Split job-specific configuration values into AzureSearchJobConfiguration (#440)
