+++ src/NuGet.Jobs.Db2AzureSearch/Job.cs
@@ -35,6 +35,7 @@ namespace NuGet.Jobs
             services.AddAzureSearch();
 
             services.Configure<Db2AzureSearchConfiguration>(configurationRoot.GetSection(ConfigurationSectionName));
+            services.Configure<AzureSearchJobConfiguration>(configurationRoot.GetSection(ConfigurationSectionName));
             services.Configure<AzureSearchConfiguration>(configurationRoot.GetSection(ConfigurationSectionName));
             services.AddTransient<Db2AzureSearchCommand>();
         }
