+++ src/NuGet.Services.AzureSearch/Catalog2AzureSearch/CatalogLeafFetcher.cs
@@ -35,7 +35,7 @@ namespace NuGet.Services.AzureSearch.Catalog2AzureSearch
             if (_options.Value.MaxConcurrentBatches <= 0)
             {
                 throw new ArgumentException(
-                    $"The {nameof(AzureSearchConfiguration.MaxConcurrentBatches)} must be greater than zero.",
+                    $"The {nameof(AzureSearchJobConfiguration.MaxConcurrentBatches)} must be greater than zero.",
                     nameof(options));
             }
         }
